practice/pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = ""
authors = [
    {name = "youngchan-itall",email = "michael.yc.cho@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Existing dependencies
    "py7zr (>=1.0.0,<2.0.0)",
    "diffusers (>=0.34.0,<0.35.0)",
    "torch (>=2.7.1,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "yt-dlp (>=2025.6.30,<2026.0.0)",
    "pymupdf (>=1.26.3,<2.0.0)",
    
    # PDF OCR processing dependencies
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "langdetect>=1.0.9",
    
    # Web API dependencies
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    
    # CLI and UI dependencies
    "rich>=13.0.0",
    "colorama>=0.4.6",
    "click>=8.1.7",
    
    # Async and concurrency
    "aiofiles>=23.0.0",
    "asyncio-mqtt>=0.13.0",
    
    # Data processing and caching
    "psutil>=5.9.0",
    "redis>=5.0.0",
    "diskcache>=5.6.0",
    
    # Configuration and logging
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    
    # Testing and development (optional)
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
